Rebases DashMapActivity to be purely a Dashboard Activity
